+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - `pctx add` now accepts `--header` and `--bearer` to add authentication without interaction
-- typescript type checking loosened for poorly documented MCPs
+- add typescript type check runtime capabilities including more typical string/array utils
 - tool descriptions updated for consistent behavior